requirements.txt
altair
pandas
streamlit
st-annotated-text
yargy
pymorphy2